Split out models and printer into their own files
