+++ Pipfile
@@ -8,7 +8,8 @@ numpy = "*"
 ipdb = "==0.13.2"
 ipython = "*"
 mypy = "*"
-black = "*"
+moonshade = {editable = true, path = "."}
+black = "==19.10b0"
 
 [requires]
 python_version = "3.8"
