+++ README.md
@@ -5,4 +5,4 @@ Photosynthesis in Python
 
 1. Install Python and Pipenv
 2. pipenv install
-3. pipenv run python src/game.py
+3. pipenv run python moonshade/game.py
