+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+
+setup(
+    name="moonshade",
+    version="0.1",
+    description="Photosynthesis in Python",
+    packages=["moonshade"],
+    package_dir={"": "."},
+)
